update FarcalculatorService in relation with sonarcloud
cast one operands of integer division to double was a bug
in FareCalculatorService and FareCalculatorServiceTest
exemple (double) 30/60;
correction by change 30 by 30f and suppression of cast double

+++ src/main/java/com/parkit/parkingsystem/service/FareCalculatorService.java
@@ -42,7 +42,7 @@ public class FareCalculatorService {
     // else we loose the decimal numbers)
 
     double durationNumberHours =
-        (duration >= (0.5 * 60 * 60 * 1000)) ? (double) (duration / (1000 * 60)) / 60 : 0;
+        (duration >= (0.5 * 60 * 60 * 1000)) ? (duration / (1000f * 60)) / 60 : 0;
 
     switch (ticket.getParkingSpot().getParkingType()) {
       case CAR: {
